requirements.txt
flask
pyjwt
mysql-connector-python
requests
typing_extensions
